+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Configuration module
+- HTTP server factory with reconnect feature
 
 ## [0.2.0] - 2021-06-24